fix(web): address all ce:review findings in nativewindui web shims
P1 fixes:
- Toggle: add prop-bridging wrapper (value/onValueChange → checked/onCheckedChange)
  to fix silent runtime bug where shadcn Switch ignored native prop names
- TextField: guard secureTextEntry+multiline combination to prevent plaintext
  password rendering in <textarea>
- Extract useColorScheme + color tokens to use-color-scheme.web.ts, breaking
  circular import theme-toggle → index.web → theme-toggle

P2 fixes:
- Export ContextMenuMethods as stub type (was missing; hidden by tsconfig exclude)
- Remove duplicate errorMessage property in TextFieldProps
- createContextSubMenu/createDropdownSubMenu now return new spread objects
  instead of mutating the passed subMenu argument
- Move @packrat/web-ui from devDependencies to dependencies in packages/ui
- Wire onSubmitEditing to onKeyDown Enter in SearchInput
- Remove unsafe `rest as Partial<ShadcnButtonProps>` spread in Button
- Remove explicit LucideIcon type annotation on ICON_MAP lookup to fix
  noUncheckedIndexedAccess conflict

P3 fixes:
- Extract shared cn utility to cn.web.ts; all shim files import from there
- AlertMethods changed from runtime value {} to export type Record<string,never>
- createContextItem/createDropdownItem changed to generic identity <T>(item:T):T
- ActivityIndicator element corrected to <output> (semantic role="status")
- picker.web.tsx collectItems uses React.isValidElement + child.type guard
- tsconfig.json excludes *.web.ts alongside *.web.tsx
- Removed unused ShadcnButtonProps import from button.web.tsx
