+++ packages/ui/nativewindui/list.web.tsx
@@ -1,10 +1,6 @@
-import { type ClassValue, clsx } from 'clsx';
 import type * as React from 'react';
-import { twMerge } from 'tailwind-merge';
 
-function cn(...inputs: ClassValue[]) {
-  return twMerge(clsx(inputs));
-}
+import { cn } from './cn.web';
 
 export interface ListDataItem {
   title: string;
