+++ packages/ui/package.json
@@ -3,9 +3,7 @@
   "version": "2.0.25",
   "private": true,
   "dependencies": {
-    "@packrat-ai/nativewindui": "^2.0.5"
-  },
-  "devDependencies": {
+    "@packrat-ai/nativewindui": "^2.0.5",
     "@packrat/web-ui": "workspace:*"
   }
 }
